+++ src/OrderScrapper/OrderScrapper/Views/ProgressView.swift
@@ -12,6 +12,9 @@ class ProgressView: UIView {
     @IBOutlet weak var headerLabel: UILabel!
     @IBOutlet weak var stepLabel: UILabel!
     @IBOutlet weak var progressView: HorizontalProgressBar!
+    @IBOutlet weak var scrapePercentage: UILabel!
+    @IBOutlet weak var cancelScrapeBtn: UIButton!
+    var buttonClickHandler: (() -> Void)?
     
     var headerText: String {
         get {
@@ -40,6 +43,23 @@ class ProgressView: UIView {
         }
     }
     
+    var scrapePercentValue: String {
+        get {
+            return ""
+        }
+        set {
+            scrapePercentage.text = newValue
+        }
+    }
+    
+    var hideCancelScrapeBtn: Bool {
+        get {
+            false
+        }
+        set {
+            cancelScrapeBtn.isHidden = newValue
+        }
+    }
     override init(frame: CGRect) {
         super.init(frame: frame)
         initView()
@@ -56,5 +76,11 @@ class ProgressView: UIView {
 
         self.contentView.frame = self.bounds
         self.addSubview(self.contentView)
+    }
+    
+    @IBAction func onCancelClick(_ sender: Any) {
+        if let clickHandler = buttonClickHandler {
+            clickHandler()
+        }
     }
 }